+++ src/cli/mc.ts
@@ -2,19 +2,19 @@
 // Copyright (C), Siemens AG 2017
 import * as program from "commander";
 import agentStatusCommand from "./commands/mc-agent-status";
+import agentTokenCommand from "./commands/mc-agent-token";
 import createEventCommand from "./commands/mc-create-event";
 import getDiagnosticCommand from "./commands/mc-get-diagnostic";
 import onboardCommand from "./commands/mc-onboard";
 import registerDiagnosticCommand from "./commands/mc-register-diagnostic";
 import serviceCredentialsCommand from "./commands/mc-service-credentials";
+import serviceTokenCommand from "./commands/mc-service-token";
 import starterJsCommand from "./commands/mc-starter-js";
 import starterTsCommand from "./commands/mc-starter-ts";
 import unregisterDiagnoticCommand from "./commands/mc-unregister-diagnostic";
 import uploadFileCommand from "./commands/mc-upload-file";
 import uploadTimeSeriesCommand from "./commands/mc-upload-timeseries";
 import versionAndHelp from "./commands/mc-version-help";
-import agentTokenCommand from "./commands/mc-agent-token";
-
 
 // * generic commands
 versionAndHelp(program);
@@ -29,6 +29,7 @@ agentStatusCommand(program);
 
 // * setup commands
 serviceCredentialsCommand(program);
+serviceTokenCommand(program);
 registerDiagnosticCommand(program);
 getDiagnosticCommand(program);
 unregisterDiagnoticCommand(program);