+++ src/cli/commands/mc-upload-timeseries.ts
@@ -15,6 +15,7 @@ import {
     retrylog,
     verboseLog
 } from "../../api/utils";
+import { displayCsvHelp } from "./command-utils";
 const mime = require("mime-types");
 
 export default (program: CommanderStatic) => {
@@ -27,7 +28,7 @@ export default (program: CommanderStatic) => {
             "required for agents with RSA_3072 profile. create with: openssl genrsa -out private.key 3072"
         )
         .option("-f, --file <timeseriesdata.csv>", "csv file containing the timeseries data to upload to mindsphere")
-        .option("-s, --size <size>", "max elements per http post", 200)
+        .option("-s, --size <size>", "max records per http post", 200)
         .option("-n, --no-validation", "switch validation off (only if you are sure that the timeseries upload works)")
         .option("-y, --retry <number>", "retry attempts before giving up", 3)
         .option("-v, --verbose", "verbose output")
@@ -104,42 +105,19 @@ export default (program: CommanderStatic) => {
                         options.verbose
                     );
 
-                    let i = 0;
+                    let messageCount = 0;
                     await csv()
                         .fromFile(uploadFile)
                         .subscribe(async json => {
                             data.push(json);
                             if (data.length >= maxSize) {
-                                verboseLog(
-                                    `posting timeseries message Nr. ${chalk.cyanBright(
-                                        ++i + ""
-                                    )} with ${chalk.cyanBright(data.length + "")} records `,
-                                    true
-                                );
-                                await retry(
-                                    options.retry,
-                                    () =>
-                                        agent.BulkPostData(convertToTdpArray(data), options.validation ? true : false),
-                                    300,
-                                    retrylog("TimeSeriesBulkUpload")
-                                );
+                                messageCount = await postChunk(messageCount, data, options, agent);
                                 data = [];
                             }
                         });
 
                     if (data.length >= 1) {
-                        verboseLog(
-                            `posting timeseries message Nr. ${chalk.cyanBright(++i + "")} with ${chalk.cyanBright(
-                                data.length + ""
-                            )} records `,
-                            true
-                        );
-                        await retry(
-                            options.retry,
-                            () => agent.BulkPostData(convertToTdpArray(data), options.validation ? true : false),
-                            300,
-                            retrylog("TimeSeriesBulkUpload")
-                        );
+                        messageCount = await postChunk(messageCount, data, options, agent);
                         data = [];
                     }
                 } catch (err) {
@@ -148,44 +126,23 @@ export default (program: CommanderStatic) => {
             })();
         })
         .on("--help", () => {
-            log("\n  Examples:\n");
-            log(`    mc ts -f timeseries.csv \t\t\t\t\t upload timeseries from the csv file to mindsphere `);
-            log(`    mc upload-timeseries --file timeseries.csv  --size 100  \t use http post size of 100 `);
-
-            log(`\n  ${chalk.cyanBright("Data Format:")} (use your own data point ids from mindsphere)\n`);
-            log(
-                `  timestamp, ${chalk.cyanBright("dataPointId")}, ${chalk.greenBright(
-                    "qualityCode"
-                )}, ${chalk.yellowBright("value")}`
-            );
-            log(
-                `  2018-08-23T09:18:21.809Z, ${chalk.cyanBright("DP-Temperature")} ,${chalk.greenBright(
-                    "0"
-                )}, ${chalk.yellowBright("20.34")}`
-            );
-            log(
-                `  2018-08-23T09:18:21.809Z, ${chalk.cyanBright("DP-Humidity")}, ${chalk.greenBright(
-                    "0"
-                )}, ${chalk.yellowBright("70.4")}`
-            );
-            log(
-                `  2018-08-23T09:18:21.809Z, ${chalk.cyanBright("DP-Pressure")}, ${chalk.greenBright(
-                    "0"
-                )}, ${chalk.yellowBright("1012.3")}`
-            );
-
-            log(
-                `\n  Make sure that the timestamp is in ISO format. The headers and the casing (timestamp, dataPointId) are important.`
-            );
-
-            log(`\n  ${chalk.cyanBright("Important:")}\n`);
-            log(
-                `    You have to configure the data source and data mappings in mindsphere asset manager before you can upload the data`
-            );
-            log(
-                `    See also: ${chalk.cyanBright(
-                    "https://documentation.mindsphere.io/resources/html/getting-connected/en-US/index.html"
-                )}`
-            );
+            displayCsvHelp(chalk.cyanBright);
         });
 };
+
+async function postChunk(messageCount: number, data: any[], options: any, agent: MindConnectAgent) {
+    verboseLog(
+        `posting timeseries message Nr. ${chalk.cyanBright(++messageCount + "")} with ${chalk.cyanBright(
+            data.length + ""
+        )} records `,
+        true
+    );
+    const tdpArray = convertToTdpArray(data);
+    await retry(
+        options.retry,
+        () => agent.BulkPostData(tdpArray, options.validation ? true : false),
+        300,
+        retrylog("TimeSeriesBulkUpload")
+    );
+    return messageCount;
+}